Type and Function Reference

Type and Function Reference

KeplerGLMap and rendering

Map aesthetics

Loading and saving/exporting

KeplerGLData

KeplerGLMap and rendering

KeplerGL.KeplerGLMap — Type

mutable struct KeplerGLMap 
    datasets::Vector{AbstractKeplerGLData}
    config::Dict{Symbol, Any}
    info::Dict{Symbol, Any}
    window::Dict{Symbol, Any}
end

Type that contains all the information to render a map. config and info are of the same structure as the map JSON serialization that Kepler.gl uses internally. datasets are instances of AbstractKeplerGLData that contain the data that the layers draw from. window is additional information not used by Kepler.gl that contains information on window size, whether the map should be centered, and so on.

KeplerGL.render — Function

render(map::KeplerGLMap)

Renders a KeplerGLMap in a new Blink.jl window and returns this window.

Required Arguments

m::KeplerGLMap: the map that should be rendered

KeplerGL.get_html — Function

get_html(m::KeplerGLMap)

Gets the html + js code to render the map in Blink

Required Arguments

m::KeplerGLMap: the map to get the HTML code for.

KeplerGL.make_dispatch_code — Function

make_dispatch_code(m::KeplerGLMap)

This function makes the dispatch code for m. It is called by render() but is sometimes also useful to call directly for debugging purposes

Required Arguments

m::KeplerGLMap: the map of which the dispatch code should be returned.

Map aesthetics

KeplerGL.hide_buttons! — Function

hide_buttons!(m::KeplerGLMap)

Hides all buttons (legend, 3d, etc) on the top right corner of a KeplerGLMap

Required Arguments

m::KeplerGLMap: the map that should be affected

KeplerGL.hide_legend! — Function

hide_legend!(m::KeplerGLMap)

Hides the legend on the top right corner of a KeplerGLMap

Required Arguments

m::KeplerGLMap: the map that should be affected

Loading and saving/exporting

KeplerGL.load_map_from_json — Function

load_map_from_json(token::String, json_map_file::String)

Returns a KeplerGLMap whose data and config has been loaded from json_map_file.

Required Arguments

token::String: mapbox token to use with the map.

json_map_file::String: the path to the Kepler.gl map JSON file which should be loaded.

KeplerGL.load_map_from_json! — Function

load_map_from_json!(m::KeplerGLMap, json_map_file::String)

Loads a kepler.gl map file from json_map_file into the map m. Overwrites any existing map config and datasets.

Required Arguments

m::KeplerGLMap: the map to which the config and data should be applied.

json_map_file::String: the path to the kepler.gl JSON map file.

KeplerGL.load_config_from_json! — Function

load_config_from_json!(m::KeplerGLMap, json_map_file::String)

Loads the config from a json file json_map_file and applies it to the map m.

Required Arguments

m::KeplerGLMap: the map to which the config should be applied.

json_map_file::String: the path to the JSON file from which the config should be loaded.

KeplerGL.get_json — Function

get_json(m::KeplerGLMap)

Returns the map in JSON format, which is also the format which Kepler.gl uses intrinsically to save map files.

Required Arguments

m::KeplerGLMap: the map of which the JSON code should be returned.

KeplerGL.export_image — Function

export_image(window,filename; 
    width = 1000, height = 800,
    imageDataUri = "data:image/png;base64,2i3u")

Exports an image from window to filename of the format specified by imageDataUri.

Required Arguments

window: the window whose current state should be exported. This is returned by render().

filename: path to the file to be saved.

width: width of the exported image, in pixels.

height: height of the exported image, in pixels.

imageDataUri = "data:image/png;base64,2i3u": export image data uri. For JPG, choose "data:image/jpg;base64,".

KeplerGLData

KeplerGL.AbstractKeplerGLData — Type

abstract type AbstractKeplerGLData end

Abstract type for datasets.

KeplerGL.FieldsRowsData — Type

struct FieldsRowsData <: AbstractKeplerGLData
    id::String
    json::String
end

Concrete type for AbstractKeplerGLData, which contains data in the fields-rows format that Kepler.gl uses natively. Used for data that is loaded from KeplerGL maps in JSON format.

KeplerGL.GeoJSONData — Type

struct GeoJSONData <: AbstractKeplerGLData
    id::String
    json::String
end

Concrete type for AbstractKeplerGLData, which contains data in GeoJSON format.

KeplerGL.CSVData — Type

struct CSVData <: AbstractKeplerGLData
    id::String
    csvstring::String
end

Concrete type for AbstractKeplerGLData, which contains data in CSV format. Used also when bringing data in via a DataFrame.